Don't do all signatures from the entire program
For each package, we now take into account the func signatures found in
itself and its direct dependencies only. This turns out to work quite
well, and means that the tool will work equally well independently of
how many packages it is given as arguments.

Previously, if any of the packages declared tons of func signatures,
they would hide matching funcs with warnings in other packages even if
the two packages weren't related.