check: remove "result X is always parameter Y"
It was rarely useful, and it was particularly annoying with patterns
where funcs would return the same parameters they got, which would make
for simpler statements and expression when chaining these funcs or when
using them as return arguments.

See the discussion in #18.

Shaves off 19 warnings off std cmd, which were mostly false positives
like the above.