check: fix "result X is always nil" warnings
They weren't working, as we used nil to mark when a return value wasn't
consistent. However, constant.Value represents a nil value as simply an
untyped nil, so the two collided and we discarded those.

Add a level of indirection so that we can tell the two nil cases apart.

Updates #21.