check: skip more inbound calls generated by ssa
The one that was giving uss trouble now is when a method declared on a
non-pointer receiver is called on a value of a pointer type.

In that case, ssa generates a call like (*T).M -> (T).M. That's not a
real call though, so don't take it into account when looking for
forwarding returns.

Fixes #21.