check: simplify code via DeleteSyntheticNodes
We had multiple pieces of code to deal with ssa's generated calls, such
as the one needed when a non-pointer method is called on a pointer
receiver.

Instead, use DeleteSyntheticNodes, which gets rid of those and "inlines"
these now removed edges. That means we no longer have to worry about
these edge cases.

The proof is that the tests still pass, and there are no changes to the
warnings emitted for all of std cmd.