check: fix occasional issues with test packages
declCounts only gets a package directory and name, and was using the
absolute directory as a key to cache its results. However, this is wrong
- if a package also has a test package, the two have the same absolute
directory containing their sources.

With the added testdata file, the issue can be seen roughly one in ten
times:

	$ go test -count 10
	--- FAIL: TestCHA (2.40s)
	    check_test.go:46: Unexpected output:
		+testdata/tags_main.go:14:34: (FooType).multImplsMethod - f2 is unused
		+testdata/tags_main.go:19:36: (*FooType).multImplsMethod2 - f3 is unused
		+testdata/tags_main.go:9:20: multipleImpls - f is unused

To fix this, also include the package name in the map key, to avoid the
conflicts.

Found while using unparam on std, where the "os" package saw this issue
quite often.