check: remove a need for interface pointers
Make use of go/constant's MakeUnknown as a special value to mark when we
didn't see the same constant in all cases. This way, we don't need to
have all the elements of the slice as pointers, removing some jumping
around memory.

We're not interested in unknown constant values anyway, so it doesn't
matter that we don't support them in this particular check.