check: don't skip warnings for closure functions
We were being overly conservative. Found thanks to a false negative in
Go's cmd/trace, and added a regression test.

Also expanded requiredViaCall to be more explicit in what instructinos
are safe or not, and why. Switch to a blacklist instead of a whitelist,
to more easily spot false negatives in the future.

ssa.Phi was the only non-safe instruction that didn't have a regression
test, so add one.

It's not yet clear why ssa.UnOp is not safe. Leave it blacklisted, for
now.