check: detect repeated args to interface parameters
We were ignoring constants behind ssa.MakeInterface, so we were ignoring
cases such as always passing the same integer constant as an interface{}
argument.

However, this requires that we start checking that the type is always
the same, not just the value. For example, code may use typed nils to
pass around types, in which case the constants are not redundant.

This also lets us get rid of unknownConst, as the nil *ssa.Const serves
as a valid zero value which is different from an untyped nil constant.