check: revert to using packages.LoadSyntax
As Alan Donovan points out, LoadAllSyntax is not required after all if
one only needs the syntax, types, and SSA for the direct set of
packages.

That's exactly our use case, so use that. This means we can avoid
re-loading all the dependencies from source each time, which can greatly
speed up the tool on large projects.