don't flag funcs used as fields
These very often lead to false positives, as fields tend to have a
function signature type which is hard to change. Moreover, it's hard to
track where the field will end up being used.

The callgraph package isn't of much help here, so implement this with
go/ssa only.

Fixes #35.