never warn on funcs used as interfaces
These often lead to false positives, as the interface value is often
called via type switches, reflection, or other methods.

As before, don't rely on callgraph for this. It doesn't deal well with
that kind of runtime type conversion.

We had a few test cases which were "correct positives", but looking at
them again, they don't seem particularly useful or stable. Better to be
on the safe side and not warn on those.

Fixes #23.