remove the -tags flag
go/packages does not support loading with a number of tags; at least not
right now. The flag did nothing, so remove it for now.

A likely solution for those requiring the flag:

	GOFLAGS=-tags=foo,bar unparam ./...

This commit is technically a breaking change, but the flag had been
broken for months. And I presume that all users are pinning a specific
version of unparam.