work around confusing crash in std
Seems to have started happening since I pulled the last Go master
version.

Nothing seems obviously wrong with my code, and it only happens 50% of
the time. Might be a bug in the ssa or callgraph packages.

A very weird corner case though, so for now, just work around the panic.