avoid false positives with implements assertions
It's common to write global declarations like:

	var _ SomeInterface = (*SomeNamed)(nil)

This way, one ensures that SomeNamed will always implement SomeInterface
in the future. However, unparam could still suggest changes to the
methods in SomeNamed, which could break this assertion.

For now, just skip the entire named type in that case. In the future, we
can only skip those methods which are necessary to implement the
interface.