work properly with method wrapper functions
When we see an instruction using a method wrapper, we are interested in
the method being wrapped, not the wrapper itself. Make findFunction
fetch the function being wrapped.

This fixes the false positive in the added test case. Before, we didn't
see that FooType.checkRedirect was being used as a field, as the Store
instruction only used the method wrapper from the method value line
above.