rewrite "called in return" code without callgraph
Another big chunk of unparam that can be implemented in pure go/ssa. We
only have one piece left.

This one was tricky, as it required keeping track of which parts of
signatures cannot be changed in a more precise way.

In the process, we uncovered some new false positives due to the removal
of callgraph. Those were fixed with go/ssa, and more regression tests
were added.

On the plus side, the newer code is more precise when it comes to false
negatives, so five more correct warnings are now detected in the main Go
repository.