check: replace last use of callgraph
Both the "result is never used" and "parameter always receives const"
checks only worked for local calls really, as we can't know how
dependent packages call exported functions of the current package.

As such, we don't really need a deep callgraph. Looking at the direct
static calls is practically the same. I was still expecting to see a few
new false positives in the testing corpus, but none have appeared.

For now, leave the new code as simple as it is. We can make it smarter
if we do find false positives.