initial move to go/analysis
The tool is largely the same. The only difference is that we no longer
control the command-line interface; this means that the output contains
absolute paths, is printed to stderr, and isn't sorted. Which is fine.

We are also disabling the debug flag for now, as it conflicts with a
flag from go/analysis.

All the tests still pass, but this probably needs refining a bit more
before we can push to master.