add preliminary support for generic code
Note that we don't yet support reporting unused parameters on generic
functions, as showed by the TODO in typeparams.txt, but at least we
don't panic as quickly as before.

Also note that running unparam on std currently panics as well,
presumably due to a bug in go/ssa that I will report shortly.

Thanks to Ainar Garipov for reporting and Ludovic Fernandez for
suggesting a fix.

Note that we need to use x/exp/typeparams to keep support for Go 1.17.x.

Fixes #62.